requirements.txt
runpod==1.7.0
numpy==1.26.4
Pillow==10.4.0
transformers==4.44.2
accelerate==0.33.0
cloudinary==1.41.0
requests==2.32.3
python-dotenv==1.0.1